+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 IMPROVEMENTS:
 
-* config: Allow setting `security_opt` option. 
+
+* config: Allow setting `security_opt` option. [[GH-382](https://github.com/hashicorp/nomad-driver-podman/pull/382)]
+* config: Add `socket` stanza to allow multiple Podman sockets to be used. [[GH-371](https://github.com/hashicorp/nomad-driver-podman/pull/371)]
 
 ## 0.6.1 (July 15, 2024)